Update gradle so that the functional task runs the integration tests
