+++ src/integrationTest/java/uk/gov/hmcts/reform/divorce/casemaintenanceservice/controller/CcdUpdateITest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.divorce.casemaintenanceservice.functionaltest;
+package uk.gov.hmcts.reform.divorce.casemaintenanceservice.controller;
 
 import com.github.tomakehurst.wiremock.matching.EqualToPattern;
 import feign.FeignException;
