+++ src/integrationTest/java/uk/gov/hmcts/reform/divorce/casemaintenanceservice/controller/RetrieveAllDraftsServiceITest.java
@@ -1,0 +1,118 @@
+package uk.gov.hmcts.reform.divorce.casemaintenanceservice.controller;
+
+import com.github.tomakehurst.wiremock.matching.EqualToPattern;
+import com.github.tomakehurst.wiremock.matching.StringValuePattern;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.test.annotation.DirtiesContext;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.TestPropertySource;
+import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+import org.springframework.web.client.HttpClientErrorException;
+import uk.gov.hmcts.reform.divorce.casemaintenanceservice.CaseMaintenanceServiceApplication;
+import uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.DraftStoreClient;
+import uk.gov.hmcts.reform.divorce.casemaintenanceservice.draftstore.model.Draft;
+import uk.gov.hmcts.reform.divorce.casemaintenanceservice.draftstore.model.DraftList;
+
+import java.util.Collections;
+
+import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
+import static com.github.tomakehurst.wiremock.client.WireMock.get;
+import static org.hamcrest.CoreMatchers.containsString;
+import static org.mockito.Mockito.when;
+import static org.springframework.http.HttpHeaders.CONTENT_TYPE;
+import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.TestConstants.TEST_DRAFT_DOCUMENT_TYPE_CCD_FORMAT;
+import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.TestConstants.TEST_SERVICE_TOKEN;
+
+@RunWith(SpringRunner.class)
+@ContextConfiguration(classes = CaseMaintenanceServiceApplication.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@PropertySource(value = "classpath:application.yml")
+@TestPropertySource(properties = {
+    "feign.hystrix.enabled=false",
+    "eureka.client.enabled=false"
+    })
+@AutoConfigureMockMvc
+@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_CLASS)
+public class RetrieveAllDraftsServiceITest extends MockSupport {
+    private static final String API_URL = "/casemaintenance/version/1/drafts";
+    private static final String DRAFTS_CONTEXT_PATH = "/drafts";
+    private static final String DRAFT_ID = "1";
+    private static final Draft DRAFT = new Draft(DRAFT_ID, null, TEST_DRAFT_DOCUMENT_TYPE_CCD_FORMAT);
+
+    @Autowired
+    private MockMvc webClient;
+
+    @Test
+    public void givenJWTTokenIsNull_whenRetrieveAllDrafts_thenReturnBadRequest() throws Exception {
+        webClient.perform(MockMvcRequestBuilders.get(API_URL))
+            .andExpect(status().isBadRequest());
+    }
+
+    @Test
+    public void givenInvalidUserToken_whenRetrieveAllDrafts_thenReturnForbiddenError() throws Exception {
+        final String message = "some message";
+        stubUserDetailsEndpoint(HttpStatus.FORBIDDEN, new EqualToPattern(USER_TOKEN), message);
+
+        webClient.perform(MockMvcRequestBuilders.get(API_URL)
+            .header(HttpHeaders.AUTHORIZATION, USER_TOKEN)
+            .accept(MediaType.APPLICATION_JSON))
+            .andExpect(status().isForbidden())
+            .andExpect(content().string(containsString(message)));
+    }
+
+    @Test
+    public void givenCouldNotConnectToAuthService_whenRetrieveAllDrafts_thenReturnHttp503() throws Exception {
+        final String message = getUserDetails();
+
+        when(serviceTokenGenerator.generate()).thenThrow(new HttpClientErrorException(HttpStatus.SERVICE_UNAVAILABLE));
+
+        stubUserDetailsEndpoint(HttpStatus.OK, new EqualToPattern(USER_TOKEN), message);
+
+        webClient.perform(MockMvcRequestBuilders.get(API_URL)
+            .header(HttpHeaders.AUTHORIZATION, USER_TOKEN))
+            .andExpect(status().isServiceUnavailable());
+    }
+
+    @Test
+    public void givenThereIsDrafts_whenRetrieveAllDrafts_thenReturnHttp200() throws Exception {
+        final String message = getUserDetails();
+
+        final DraftList draftList = new DraftList(Collections.singletonList(DRAFT), null);
+
+        when(serviceTokenGenerator.generate()).thenReturn(TEST_SERVICE_TOKEN);
+
+        stubGetDraftEndpoint(new EqualToPattern(USER_TOKEN), new EqualToPattern(TEST_SERVICE_TOKEN),
+            ObjectMapperTestUtil.convertObjectToJsonString(draftList));
+
+        stubUserDetailsEndpoint(HttpStatus.OK, new EqualToPattern(USER_TOKEN), message);
+
+        webClient.perform(MockMvcRequestBuilders.get(API_URL)
+            .header(HttpHeaders.AUTHORIZATION, USER_TOKEN))
+            .andExpect(status().isOk())
+            .andExpect(content().json(ObjectMapperTestUtil.convertObjectToJsonString(draftList)));
+    }
+
+    private void stubGetDraftEndpoint(StringValuePattern authHeader, StringValuePattern serviceToken, String message) {
+        draftStoreServer.stubFor(get(DRAFTS_CONTEXT_PATH)
+            .withHeader(HttpHeaders.AUTHORIZATION, authHeader)
+            .withHeader(DraftStoreClient.SERVICE_AUTHORIZATION_HEADER_NAME, serviceToken)
+            .withHeader(DraftStoreClient.SECRET_HEADER_NAME, new EqualToPattern(ENCRYPTED_USER_ID))
+            .willReturn(aResponse()
+                .withStatus(HttpStatus.OK.value())
+                .withHeader(CONTENT_TYPE, APPLICATION_JSON_VALUE)
+                .withBody(message)));
+    }
+}
