+++ src/integrationTest/java/uk/gov/hmcts/reform/divorce/casemaintenanceservice/controller/SearchCasesITest.java
@@ -1,6 +1,8 @@
-package uk.gov.hmcts.reform.divorce.casemaintenanceservice.functionaltest;
+package uk.gov.hmcts.reform.divorce.casemaintenanceservice.controller;
 
 import com.github.tomakehurst.wiremock.matching.EqualToPattern;
+import com.google.common.collect.ImmutableMap;
+import feign.FeignException;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,12 +22,18 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
+import uk.gov.hmcts.reform.ccd.client.model.SearchResult;
 import uk.gov.hmcts.reform.divorce.casemaintenanceservice.CaseMaintenanceServiceApplication;
+import uk.gov.hmcts.reform.divorce.casemaintenanceservice.draftstore.domain.model.CitizenCaseState;
+
+import java.util.Arrays;
 
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.TestConstants.TEST_SERVICE_TOKEN;
+import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.TestConstants.TEST_USER_EMAIL;
+import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.domain.model.CcdCaseProperties.D8_PETITIONER_EMAIL;
 
 @RunWith(SpringRunner.class)
 @ContextConfiguration(classes = CaseMaintenanceServiceApplication.class)
@@ -37,8 +45,9 @@ import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.TestConstants.T
     })
 @AutoConfigureMockMvc
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_CLASS)
-public class RetrievePetitionByCaseIdITest extends MockSupport {
-    private static final String API_URL = "/casemaintenance/version/1/case";
+public class SearchCasesITest extends MockSupport {
+    private static final String API_URL = "/casemaintenance/version/1/search";
+
 
     @Value("${ccd.jurisdictionid}")
     private String jurisdictionId;
@@ -53,76 +62,60 @@ public class RetrievePetitionByCaseIdITest extends MockSupport {
     private MockMvc webClient;
 
     @Test
-    public void givenCaseExistsWithId_whenRetrievePetitionById_thenReturnCaseDetails()
-        throws Exception {
+    public void whenSearchCases_thenReturnCcdResult() throws Exception {
         final String message = getUserDetails();
+
+        final CaseDetails caseDetails2 = createCaseDetails(1L, CitizenCaseState.ISSUED.getValue());
+        final CaseDetails caseDetails3 = createCaseDetails(2L, CitizenCaseState.PENDING_REJECTION.getValue());
+        final CaseDetails caseDetails4 = createCaseDetails(3L, CitizenCaseState.PENDING_REJECTION.getValue());
+
+        String query = "{}";
         stubUserDetailsEndpoint(HttpStatus.OK, new EqualToPattern(USER_TOKEN), message);
-
-        final Long caseId = 1L;
-        final String testCaseId = String.valueOf(caseId);
-
-        final CaseDetails caseDetails = createCaseDetails(caseId, "state");
+        SearchResult expectedResult = SearchResult.builder()
+            .cases(Arrays.asList(caseDetails2, caseDetails3, caseDetails4))
+            .total(10)
+            .build();
 
         when(serviceTokenGenerator.generate()).thenReturn(TEST_SERVICE_TOKEN);
         when(coreCaseDataApi
-            .readForCitizen(USER_TOKEN, TEST_SERVICE_TOKEN, USER_ID, jurisdictionId, caseType, testCaseId))
-            .thenReturn(caseDetails);
+            .searchCases(USER_TOKEN, TEST_SERVICE_TOKEN, caseType, query))
+            .thenReturn(expectedResult);
 
-        webClient.perform(MockMvcRequestBuilders.get(API_URL + "/" + testCaseId)
+        webClient.perform(MockMvcRequestBuilders.post(API_URL)
+            .content(query)
             .header(HttpHeaders.AUTHORIZATION, USER_TOKEN)
             .accept(MediaType.APPLICATION_JSON))
             .andExpect(status().isOk())
             .andExpect(content()
                 .json(ObjectMapperTestUtil
-                    .convertObjectToJsonString(caseDetails)));
+                    .convertObjectToJsonString(expectedResult)));
     }
 
     @Test
-    public void givenCaseExistsWithId_whenRetrievePetitionByIdWithCaseworker_thenReturnCaseDetails()
-        throws Exception {
-        final String message = getCaseWorkerUserDetails();
-        stubUserDetailsEndpoint(HttpStatus.OK, new EqualToPattern(BEARER_CASE_WORKER_TOKEN), message);
-
-        final Long caseId = 1L;
-        final String testCaseId = String.valueOf(caseId);
-
-        final CaseDetails caseDetails = createCaseDetails(caseId, "state");
-
-        when(serviceTokenGenerator.generate()).thenReturn(TEST_SERVICE_TOKEN);
-        when(coreCaseDataApi
-            .readForCaseWorker(BEARER_CASE_WORKER_TOKEN, TEST_SERVICE_TOKEN, CASE_WORKER_USER_ID,
-                jurisdictionId, caseType, testCaseId))
-            .thenReturn(caseDetails);
-
-        webClient.perform(MockMvcRequestBuilders.get(API_URL + "/" + testCaseId)
-            .header(HttpHeaders.AUTHORIZATION, BEARER_CASE_WORKER_TOKEN)
-            .accept(MediaType.APPLICATION_JSON))
-            .andExpect(status().isOk())
-            .andExpect(content()
-                .json(ObjectMapperTestUtil
-                    .convertObjectToJsonString(caseDetails)));
-    }
-
-    @Test
-    public void givenCaseDoesNotExistWithId_whenRetrievePetitionById_thenReturnNotFound() throws Exception {
+    public void givenCcdError_whenSearchCases_thenPropagateCcdError() throws Exception {
         final String message = getUserDetails();
         stubUserDetailsEndpoint(HttpStatus.OK, new EqualToPattern(USER_TOKEN), message);
 
-        final Long caseId = 1L;
-        final String testCaseId = String.valueOf(caseId);
+        String query = "{}";
 
         when(serviceTokenGenerator.generate()).thenReturn(TEST_SERVICE_TOKEN);
         when(coreCaseDataApi
-            .readForCitizen(USER_TOKEN, TEST_SERVICE_TOKEN, USER_ID, jurisdictionId, caseType, testCaseId))
-            .thenReturn(null);
+            .searchCases(USER_TOKEN, TEST_SERVICE_TOKEN, caseType, query))
+            .thenThrow(new FeignException.BadRequest("Malformed url", query.getBytes()));
 
-        webClient.perform(MockMvcRequestBuilders.get(API_URL + "/" + testCaseId)
+        webClient.perform(MockMvcRequestBuilders.post(API_URL)
+            .content(query)
             .header(HttpHeaders.AUTHORIZATION, USER_TOKEN)
             .accept(MediaType.APPLICATION_JSON))
-            .andExpect(status().isNotFound());
+            .andExpect(status().is4xxClientError())
+            .andExpect(content().string("Malformed url - {}"));
     }
 
     private CaseDetails createCaseDetails(Long id, String state) {
-        return CaseDetails.builder().id(id).state(state).build();
+        return CaseDetails.builder()
+            .id(id)
+            .state(state)
+            .data(ImmutableMap.of(D8_PETITIONER_EMAIL, TEST_USER_EMAIL))
+            .build();
     }
 }
