+++ src/integrationTest/resources/ccd-submission-payload/base-amended-case.json
