+++ src/integrationTest/resources/ccd-submission-payload/base-case.json
